[squash] rename file
